fix alpha mock dev mode
