+++ workspaces/topology/plugins/topology/dev/alpha/index.tsx
@@ -29,164 +29,12 @@ import {
   createFrontendModule,
   pluginHeaderActionsApiRef,
 } from '@backstage/frontend-plugin-api';
-import { appThemeApiRef, useApi } from '@backstage/core-plugin-api';
-import useObservable from 'react-use/esm/useObservable';
-import { Box, MenuItem, MenuSection, Popover } from '@backstage/ui';
-import { RiContrastLine } from '@remixicon/react';
-import { Menu } from 'react-aria-components';
-import {
-  Sidebar,
-  SidebarGroup,
-  SidebarItem,
-  SidebarScrollWrapper,
-  SidebarSpace,
-} from '@backstage/core-components';
-import { NavContentBlueprint } from '@backstage/plugin-app-react';
-import {
-  SidebarLanguageSwitcher,
-  SidebarSignOutButton,
-} from '@backstage/dev-utils';
 
 import {
   topologyCatalogModule,
   topologyTranslationsModule,
 } from '../../src/alpha';
-import {
-  cloneElement,
-  MouseEvent,
-  ReactElement,
-  ReactNode,
-  useCallback,
-  useRef,
-  useState,
-} from 'react';
-
-function ThemeMenuIcon({
-  active,
-  icon,
-}: {
-  active?: boolean;
-  icon?: ReactElement;
-}) {
-  if (icon) {
-    return cloneElement(icon, {
-      fontSize: 'small',
-      color: active ? 'primary' : undefined,
-    });
-  }
-
-  return (
-    <RiContrastLine
-      size={20}
-      color={active ? 'var(--bui-fg-accent)' : undefined}
-    />
-  );
-}
-
-function getThemeLabel(title: ReactNode, fallback: string): string {
-  return typeof title === 'string' ? title : fallback;
-}
-
-function SidebarThemeSwitcher() {
-  const appThemeApi = useApi(appThemeApiRef);
-  const themeId = useObservable(
-    appThemeApi.activeThemeId$(),
-    appThemeApi.getActiveThemeId(),
-  );
-  const themes = appThemeApi.getInstalledThemes();
-  const activeTheme = themes.find(t => t.id === themeId);
-  const anchorRef = useRef<HTMLElement | null>(null);
-  const [open, setOpen] = useState(false);
-
-  const handleSelectTheme = (newThemeId?: string) => {
-    if (newThemeId && themes.some(t => t.id === newThemeId)) {
-      appThemeApi.setActiveThemeId(newThemeId);
-    } else {
-      appThemeApi.setActiveThemeId(undefined);
-    }
-    setOpen(false);
-  };
-
-  const ActiveIcon = useCallback(
-    () => <ThemeMenuIcon icon={activeTheme?.icon} />,
-    [activeTheme],
-  );
-
-  return (
-    <>
-      <SidebarItem
-        icon={ActiveIcon}
-        text="Switch Theme"
-        onClick={(event: MouseEvent) => {
-          anchorRef.current = event.currentTarget as HTMLElement;
-          setOpen(isOpen => !isOpen);
-        }}
-      />
-      <Popover
-        triggerRef={anchorRef}
-        isOpen={open}
-        onOpenChange={setOpen}
-        placement="right bottom"
-        hideArrow
-      >
-        <Box bg="neutral">
-          <Menu>
-            <MenuSection title="Choose a theme">
-              <MenuItem
-                id="auto"
-                textValue="Auto"
-                iconStart={<ThemeMenuIcon active={themeId === undefined} />}
-                onAction={() => handleSelectTheme(undefined)}
-              >
-                Auto
-              </MenuItem>
-              {themes.map(theme => {
-                const active = theme.id === themeId;
-                const label = getThemeLabel(theme.title, theme.id);
-                return (
-                  <MenuItem
-                    key={theme.id}
-                    id={theme.id}
-                    textValue={label}
-                    iconStart={
-                      <ThemeMenuIcon icon={theme.icon} active={active} />
-                    }
-                    onAction={() => handleSelectTheme(theme.id)}
-                  >
-                    {label}
-                  </MenuItem>
-                );
-              })}
-            </MenuSection>
-          </Menu>
-        </Box>
-      </Popover>
-    </>
-  );
-}
-
-const devSidebarContent = NavContentBlueprint.make({
-  params: {
-    component: ({ navItems }) => {
-      const nav = navItems.withComponent(item => (
-        <SidebarItem icon={() => item.icon} to={item.href} text={item.title} />
-      ));
-      return (
-        <Sidebar>
-          <SidebarGroup label="Menu">
-            <SidebarScrollWrapper>
-              {nav.take('page:catalog')}
-            </SidebarScrollWrapper>
-          </SidebarGroup>
-          <SidebarSpace />
-          <SidebarThemeSwitcher />
-          <SidebarLanguageSwitcher />
-          <SidebarSignOutButton />
-        </Sidebar>
-      );
-    },
-  },
-});
+import { devSidebarContent } from './shared';
 
 const devNavModule = createFrontendModule({
   pluginId: 'app',
